update refactoring: move signature validation into a dedicated class
